api_server/requirements.txt
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==2.2.0
cryptography==2.8
Flask==1.1.1
Flask-AuditLog==1.0
Flask-Caching==1.7.0
Flask-Cors==3.0.7
Flask-No-Cache==1.0
Flask-SSLify==0.1.5
geopy==1.21.0
google-api-core==1.16.0
google-auth==1.11.0
google-cloud-core==1.3.0
google-cloud-datastore==1.9.0
google-cloud-pubsub==0.45.0
googleapis-common-protos==1.51.0
googlemaps==3.0.2
grpc-google-iam-v1==0.12.3
grpcio==1.26.0
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==2.6.0
jwkaas==0.4
MarkupSafe==1.1.1
openapi-spec-validator==0.2.8
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.6.0
pytz==2019.3
PyYAML==5.3
requests==2.22.0
rsa==4.0
six==1.14.0
swagger-ui-bundle==0.0.2
urllib3==1.25.8
Werkzeug==0.16.0
